App/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
PyJWT==2.1.0
Werkzeug==2.0.1
flask-cors==3.0.10
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
psycopg2-binary==2.9.7
flasgger==0.9.5
